requirements.txt
flask>=2.3.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
imutils>=0.5.4
scikit-image>=0.22.0
pillow>=10.0.0
python-magic-bin>=0.4.14 